FIX: auto arranging gets wrong object height
obj->bounding_box().size() is not the real object size if the object has been rotated.

jira: STUDIO-5999
(cherry picked from commit 479ea9fb02f55d24f27c94633f3d852bd5c62c83)

+++ src/slic3r/GUI/Selection.cpp
@@ -3096,7 +3096,7 @@ void Selection::paste_objects_from_clipboard()
     if (src_objects.size() > 1) {
         BoundingBoxf3 bbox_all;
         for (const ModelObject *src_object : src_objects) {
-            BoundingBoxf3 bbox = src_object->instance_convex_hull_bounding_box(0);
+            BoundingBoxf3 bbox = src_object->instance_convex_hull_bounding_box(size_t(0));
             bbox_all.merge(bbox);
         }
         auto bsize = bbox_all.size();
@@ -3112,7 +3112,7 @@ void Selection::paste_objects_from_clipboard()
         ModelObject* dst_object = m_model->add_object(*src_object);
 
         // BBS: find an empty cell to put the copied object
-        BoundingBoxf3 bbox = src_object->instance_convex_hull_bounding_box(0);
+        BoundingBoxf3 bbox = src_object->instance_convex_hull_bounding_box(size_t(0));
 
         Vec3d displacement;
         bool  in_current  = plate->intersects(bbox);
